Actions separated into separate files within actions folder
* Started Separation of files as described in #8
